1.登陆接口返回user_id 2.订单充值接口充满后自动发货 3.debug订单充值记录库存id字段太小

+++ database/migrations/2017_12_26_174518_create_order_charge_records_table.php
@@ -18,7 +18,7 @@ class CreateOrderChargeRecordsTable extends Migration
             $table->unsignedInteger('user_id')->comment('用户id');
             $table->string('order_no', 30)->comment('订单号');
             $table->unsignedInteger('game_gold')->comment('充的游戏币数');
-            $table->tinyInteger('stock_id')->comment('库存id');
+            $table->integer('stock_id')->comment('库存id');
             $table->timestamps();
 
             $table->index('user_id');
